+++ sqs/src/main/scala/monix/connect/sqs/inbound/InboundMessage.scala
@@ -1,4 +1,4 @@
-package monix.connect.sqs.domain.inbound
+package monix.connect.sqs.inbound
 
 import monix.connect.sqs.MessageAttribute
 import monix.connect.sqs.domain.QueueUrl
@@ -13,6 +13,14 @@ import scala.jdk.CollectionConverters._
   *
   * This class has a private constructor as it should not be used directly,
   * use one of its implementations [[StandardMessage]] or [[FifoMessage]].
+  *
+  * The reason for that distinction is to make the api easier and transparent
+  * to use on the user side, because different queues expect different attributes,
+  * for example a [[FifoMessage]] has `groupId` and `deduplicaitonId`, but that
+  * is not valid for standard queues.
+  * On the other hand [[StandardMessage]]s accept `delayDurations`, which are not
+  * allowed in fifo implementation.
+  * Thus it enforces the user to use the right attributes for their use case.
   *
   */
 class InboundMessage private[sqs](body: String,
