+++ sqs/src/main/scala/monix/connect/sqs/inbound/SqsParBatchSink.scala
@@ -15,12 +15,11 @@
  * limitations under the License.
  */
 
-package monix.connect.sqs
+package monix.connect.sqs.inbound
 
 import com.typesafe.scalalogging.StrictLogging
-import monix.connect.sqs.SqsParBatchSink.groupMessagesInBatches
+import monix.connect.sqs.SqsOp
 import monix.connect.sqs.domain.QueueUrl
-import monix.connect.sqs.domain.inbound.InboundMessage
 import monix.eval.Task
 import monix.execution.cancelables.AssignableCancelable
 import monix.execution.{Ack, Callback, Scheduler}
